requirements.txt
# LLM SDK
openai>=1.0.0

# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0

# Data Validation
pydantic>=2.0.0

# HTTP Client
requests>=2.25.0

# Configuration & File Formats
python-dotenv>=1.0.0
PyYAML>=6.0
